+++ app/javascript/components/training-data/Dashboard.tsx
