+++ app/javascript/components/training-data/code-tagger/left-pane/LeftPane.tsx
