+++ app/javascript/components/training-data/code-tagger/left-pane/index.ts
