+++ app/javascript/components/training-data/code-tagger/right-pane/AutomationRules.tsx
