+++ app/javascript/components/training-data/code-tagger/right-pane/PlaceholderStuff.tsx
