+++ app/javascript/components/training-data/code-tagger/right-pane/RightPane.tsx
