+++ app/javascript/components/training-data/code-tagger/right-pane/index.ts
