+++ app/javascript/components/training-data/dashboard/TaggableSolution.tsx
