+++ app/javascript/components/training-data/dashboard/TaggableSolutionList.tsx
