+++ app/views/training_data/code_tags_samples/index.html.haml
@@ -1,0 +1,4 @@
+- require_stylesheet "mentoring"
+
+#mentor-inbox-page
+  = render ViewComponents::TrainingData::Header.new
