+++ app/views/training_data/code_tags_samples/show.html.haml
@@ -1,0 +1,4 @@
+- require_stylesheet "mentoring"
+
+#page-code-tagger
+  = ReactComponents::TrainingData::CodeTagger.new
