+++ app/views/training_data/dashboard/show.html.haml
@@ -1,4 +1,4 @@
 - require_stylesheet "mentoring"
 
-#page-solution-tagger
+#page-code-tagger
   = ReactComponents::Contributing::SolutionTagger.new
